Está sendo trablhado os metodos de exibir cursos por nota e os cursos favoritos
