+++ src/main/java/sistema/modelo/dao/area/AreaDAO.java
@@ -1,9 +1,6 @@
 package sistema.modelo.dao.area;
 
-import java.util.List;
-
 import sistema.modelo.entidade.area.Area;
-import sistema.modelo.entidade.curso.Curso;
 
 public interface AreaDAO {
 	
@@ -13,5 +10,4 @@ public interface AreaDAO {
 	
 	void atualizarArea(Area area);
 	
-	List<Curso> consultaAreaCurso(Area area);
 }
