+++ src/main/java/sistema/modelo/dao/curso/CursoDAO.java
@@ -2,7 +2,10 @@ package sistema.modelo.dao.curso;
 
 import java.util.List;
 
+import sistema.modelo.entidade.aluno.Aluno;
+import sistema.modelo.entidade.area.Area;
 import sistema.modelo.entidade.curso.Curso;
+import sistema.modelo.entidade.instituicao.Instituicao;
 
 public interface CursoDAO {
 
@@ -11,5 +14,9 @@ public interface CursoDAO {
 	void atualizarCurso(Curso curso);
 	
 	List<Curso> recuperarCursos();
+	List<Curso> consultaInstituicaoCurso(Instituicao instituicao);
+	List<Curso> consultaAreaCurso(Area area);
+	List<Curso> consultaNotaCurso(Aluno aluno);
+	List<Curso> exibirCursosFavoritos(Aluno aluno);
 
 }
