+++ src/main/java/sistema/modelo/dao/curso/CursoDAOlmpl.java
@@ -4,11 +4,16 @@ import java.util.List;
 
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Join;
+import javax.persistence.criteria.ParameterExpression;
 import javax.persistence.criteria.Root;
 
 import org.hibernate.Session;
 
+import sistema.modelo.entidade.aluno.Aluno;
+import sistema.modelo.entidade.area.Area;
 import sistema.modelo.entidade.curso.Curso;
+import sistema.modelo.entidade.instituicao.Instituicao;
 import sistema.modelo.factory.conexao.FactoryConexao;
 
 public class CursoDAOlmpl implements CursoDAO {
@@ -145,5 +150,180 @@ public class CursoDAOlmpl implements CursoDAO {
 		}
 
 		return curso;
+	}
+	
+	public List<Curso> consultaInstituicaoCurso(Instituicao instituicao) {
+
+		Session sessao = null;
+		List<Curso> consultaCurso = null;
+
+		try {
+
+			sessao = banco.getConectarBanco().openSession();
+			sessao.beginTransaction();
+
+			CriteriaBuilder construtor = sessao.getCriteriaBuilder();
+
+			CriteriaQuery<Curso> criteria = construtor.createQuery(Curso.class);
+			Root<Curso> raizCurso = criteria.from(Curso.class);
+
+			Join<Curso, Instituicao> juncaoInstituicao = raizCurso.join("instituicao");
+
+			ParameterExpression<Long> idinst = construtor.parameter(Long.class);
+			criteria.where(construtor.equal(juncaoInstituicao.get("id"), idinst));
+
+			consultaCurso = sessao.createQuery(criteria).setParameter(idinst, instituicao.getId()).getResultList();
+
+			sessao.getTransaction().commit();
+
+		} catch (Exception sqlException) {
+
+			sqlException.printStackTrace();
+
+			if (sessao.getTransaction() != null) {
+				sessao.getTransaction().rollback();
+			}
+
+		} finally {
+
+			if (sessao != null) {
+				sessao.close();
+			}
+		}
+
+		return consultaCurso;
+	}
+	
+public List<Curso> consultaAreaCurso(Area area) {
+		
+		Session sessao = null;
+		List<Curso> consultaArea = null;
+
+		try {
+
+			sessao =  banco.getConectarBanco().openSession();
+			sessao.beginTransaction();
+
+			CriteriaBuilder construtor = sessao.getCriteriaBuilder();
+
+			CriteriaQuery<Curso> criteria = construtor.createQuery(Curso.class);
+			Root<Curso> raizCurso = criteria.from(Curso.class);
+
+			Join<Curso, Area> juncaoArea = raizCurso.join("area");
+
+			ParameterExpression<Long> idArea = construtor.parameter(Long.class);
+			criteria.where(construtor.equal(juncaoArea.get("idArea"), idArea));
+
+			consultaArea = sessao.createQuery(criteria).setParameter(idArea, area.getIdArea()).getResultList();
+
+			sessao.getTransaction().commit();
+
+		} catch (Exception sqlException) {
+
+			sqlException.printStackTrace();
+
+			if (sessao.getTransaction() != null) {
+				sessao.getTransaction().rollback();
+			}
+
+		} finally {
+
+			if (sessao != null) {
+				sessao.close();
+			}
+		}
+
+		return consultaArea;
+	}
+	
+	public List<Curso> consultaNotaCurso(Aluno aluno) {
+
+		Session sessao = null;
+		List<Curso> consultaNota = null;
+
+		try {
+
+			sessao =  banco.getConectarBanco().openSession();
+			sessao.beginTransaction();
+
+			CriteriaBuilder construtor = sessao.getCriteriaBuilder();
+
+			CriteriaQuery<Curso> criteria = construtor.createQuery(Curso.class);
+			Root<Curso> raizCurso = criteria.from(Curso.class);
+			
+			Join<Curso, Aluno> juncaoAluno = raizCurso.join("aluno");
+			
+			ParameterExpression<Double> notaAluno = construtor.parameter(double.class);
+			criteria.where(construtor.lessThanOrEqualTo(juncaoAluno.get("nota"), notaAluno));
+
+			consultaNota = sessao.createQuery(criteria).setParameter(notaAluno, aluno.getNota()).getResultList();
+
+			sessao.getTransaction().commit();
+
+		} catch (Exception sqlException) {
+
+			sqlException.printStackTrace();
+
+			if (sessao.getTransaction() != null) {
+				sessao.getTransaction().rollback();
+			}
+
+		} finally {
+
+			if (sessao != null) {
+				sessao.close();
+			}
+		}
+
+		return consultaNota;
+	}
+		
+		public List<Curso> exibirCursosFavoritos(Aluno aluno) {
+			
+			Session sessao = null;
+			List<Curso> favoritos = null;
+
+		try {
+
+			sessao =  banco.getConectarBanco().openSession();
+			sessao.beginTransaction();
+
+			CriteriaBuilder construtor = sessao.getCriteriaBuilder();
+
+			CriteriaQuery<Curso> criteria = construtor.createQuery(Curso.class);
+			Root<Curso> raizFavorito = criteria.from(Curso.class);
+			
+			Join<Curso, Aluno> juncaoAlunoCursoFavorito = raizFavorito.join("aluno_curso_favorito");
+			
+			ParameterExpression<Long> idcurso = construtor.parameter(Long.class);
+			criteria.where(construtor.equal(juncaoAlunoCursoFavorito.get("idCurso"), idcurso));
+					
+			Join<Curso, Aluno> juncaoAluno = raizFavorito.join("aluno");
+			
+			ParameterExpression<Long> idaluno = construtor.parameter(Long.class);
+			criteria.where(construtor.equal(juncaoAluno.get("id"), idaluno));
+			
+
+			favoritos = sessao.createQuery(criteria).setParameter(idaluno, aluno.getId()).getResultList();
+
+			sessao.getTransaction().commit();
+			
+
+		} catch (Exception sqlException) {
+
+			sqlException.printStackTrace();
+
+			if (sessao.getTransaction() != null) {
+				sessao.getTransaction().rollback();
+			}
+
+		} finally {
+
+			if (sessao != null) {
+				sessao.close();
+			}
+		}
+
+		return favoritos;
 	}
 }
