+++ src/main/java/sistema/modelo/dao/instituicao/InstituicaoDAO.java
@@ -2,7 +2,6 @@ package sistema.modelo.dao.instituicao;
 
 import java.util.List;
 
-import sistema.modelo.entidade.curso.Curso;
 import sistema.modelo.entidade.instituicao.Instituicao;
 
 public interface InstituicaoDAO {
@@ -14,7 +13,5 @@ public interface InstituicaoDAO {
 	void atualizarInstituicao(Instituicao instuicao);
 
 	List<Instituicao> recuperarInstituicoes();
-	
-	List<Curso> consultaInstituicaoCurso(Instituicao instituicao);
 	
 }
